point2f is now deprecated

+++ gleany/src/com/deweyvm/gleany/data/Recti.scala
@@ -32,7 +32,7 @@ case class Recti(x: Int, y: Int, width: Int, height: Int) {
   lazy val bottom: Int = y + height - 1
   lazy val center: Point2i = Point2i(x + width / 2, y + height / 2)
 
-  def contains(point: Point2f): Boolean =
+  def contains(point: Point2d): Boolean =
     !(point.x < x || point.x > right || point.y < y || point.y > bottom)
 
   def intersects(other: Recti): Boolean = {
